### 2023-07-10 #5645 @editorialbot
[PRE REVIEW]: pudu: A Python library for agnostic feature selection and explainability of Machine Learning classification and regression problems.
**Submitting author:** @enricgrau (<a href="http://orcid.org/0000-0002-8357-5824">Enric Grau-Luque</a>)
**Repository:** https://github.com/pudu-py/pudu
**Branch with paper.md** (empty if default branch): main
**Version:** 0.3.0
