+++ paper/paper.md
@@ -44,7 +44,7 @@ Even though there are products and libraries available for explaining different 
 
 # Overview
 
-The **pudu** library is a Python toolkit created to help make sense of machine learning results. It does this by taking a clear and deterministic approach, performing sensitivity analysis on classification and regression tasks. This toolbox uses the same basic principles as RELIEF, but offers the user plenty of flexibility with parameters and use cases. Additionally, it offers activation analysis for Neural Networks, adapting some of the guidelines defined in [Bau2018] to measure the impact of feature changes through sensitivity analysis. **pudu** measures the importance of features based on how changes in the input affect the target instance. To do this, **pudu** works under the assumption that the algorithm in use has a classification probability function or a prediction function for classification and regression problems, respectively. Using this, it changes the feature space in a predictable way and evaluates the importance of the defined features based on the changes to the target. **pudu** is versatile: it can analyze both classification and regression problems for both 1-dimensional and 2-dimensional data types as input. **pudu** creates visual explanations in an easy-to-understand way by highlighting the importance of a feature on each decision, and also offers the ability to provide localized explanations by selecting specific areas of interest.
+The **pudu** library is a Python toolkit created to help make sense of machine learning results. It does this by taking a clear and deterministic approach, performing sensitivity analysis on classification and regression tasks. This toolbox uses the same basic principles as RELIEF, but offers the user plenty of flexibility with parameters and use cases. Additionally, it offers activation analysis for Neural Networks, adapting some of the guidelines defined in [Bau2018] to measure the impact of feature changes through sensitivity analysis. **pudu** measures the importance of features based on how changes in the input affect the target instance. To do this, **pudu** works under the assumption that the algorithm in use has a classification probability function or a prediction function for classification and regression problems, respectively. Using this, it changes the feature space in a predictable way and evaluates the importance of the defined features based on the changes to the target. **pudu** is versatile: it can analyze both classification and regression problems for both 1-dimensional and 2-dimensional data types as input, such as timeseries, sppectroscopic data, and images. **pudu** creates visual explanations in an easy-to-understand way by highlighting the importance of a feature on each decision, and also offers the ability to provide localized explanations by selecting specific areas of interest.
 
 To illustrate this functionality, examples using scikit-learn [@Pedregosa2011], keras [@chollet2018keras], and localreg [@Marholm2022] are included in the documentation, along with the use of LIME and GradCAM to show how **pudu** can complement the understanding of AI decisions in different types of problems.
 
