+++ __init__.py
@@ -1,5 +1,1 @@
-"""Top-level package for pudu."""
 
-__author__ = """Enric Tomas Grau-Luque"""
-__email__ = 'patenric@gmail.com'
-__version__ = '0.3.0'
