+++ setup.py
@@ -53,6 +53,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/pudu-py/pudu',
-    version='0.3.0',
+    version='0.3.1',
     zip_safe=False,
 )
