+++ pudu/pudu.py
@@ -94,7 +94,8 @@ class pudu:
 
         # Initial values
         sh = np.array(self.x).shape
-        d_temp = np.zeros((sh[0], sh[1], sh[2], sh[3]))
+
+        self.imp = np.zeros((sh[0], sh[1], sh[2], sh[3]))
 
         scope, window, padd, evolution, total = standards.params_std(self.y, sh, scope, window, padding, evolution)
 
@@ -104,11 +105,10 @@ class pudu:
         while row <= scope[0][1] - padd[0][1] - window[0]:
             col = padd[1][0] + scope[1][0]
             while col <= scope[1][1] - padd[1][1] - window[1]:
-                x_copy = copy.deepcopy(self.x)
-
                 mask_val = mask.apply(section, total)
 
                 if mask_val == 1:
+                    x_copy = copy.deepcopy(self.x)
 
                     row_idx, col_idx = np.meshgrid(range(window[0]), range(window[1]), indexing='ij')
                     row_idx, col_idx = row_idx + row, col_idx + col
@@ -124,10 +124,11 @@ class pudu:
                         val = abs(val)
 
                     if np.shape(val):
-                        d_temp[0, row:row+window[0], col:col+window[1], 0] = val[evolution]
+                        self.imp[0, row:row+window[0], col:col+window[1], 0] = val[evolution]
                     else:
-                        d_temp[0, row:row+window[0], col:col+window[1], 0] = val
-                
+                        self.imp[0, row:row+window[0], col:col+window[1], 0] = val
+                    
+                    del x_copy, temp, temp2
                 else:
                     pass
 
@@ -135,11 +136,10 @@ class pudu:
 
                 col += window[1]
             row += window[0]
-
-        self.imp = d_temp
         
-        max_val, min_val = d_temp.max(), d_temp.min()
-        self.imp_rel = (d_temp - min_val) / (max_val - min_val)
+        max_val, min_val = self.imp.max(), self.imp.min()
+        self.imp_rel = (self.imp - min_val) / (max_val - min_val)
+
 
 
     def speed(self, window=1, scope=None, evolution=None, padding='center',
@@ -175,7 +175,7 @@ class pudu:
 
         # Initial values
         sh = np.array(self.x).shape
-        d_temp = np.zeros((sh[0], sh[1], sh[2], sh[3]))
+        self.spe = np.zeros((sh[0], sh[1], sh[2], sh[3]))
         
         scope, window, padd, evolution, total = standards.params_std(self.y, sh, scope, window, padding, evolution)
 
@@ -217,7 +217,7 @@ class pudu:
                     var_x = [i for i in range(len(p))]
                     var_y = [i for i in p]
                             
-                    d_temp[0, row:row+window[0], col:col+window[1], 0] = np.polyfit(var_x, var_y, 1)[0]
+                    self.spe[0, row:row+window[0], col:col+window[1], 0] = np.polyfit(var_x, var_y, 1)[0]
 
                 else:
                     pass    
@@ -226,11 +226,9 @@ class pudu:
 
                 col += window[1]
             row += window[0]
-
-        self.spe = d_temp
-        
-        max_val, min_val = d_temp.max(), d_temp.min()
-        self.spe_rel = (d_temp - min_val) / (max_val - min_val)
+       
+        max_val, min_val = self.spe.max(), self.spe.min()
+        self.spe_rel = (self.spe - min_val) / (max_val - min_val)
 
     
     def synergy(self, window=1, inspect=0, scope=None, absolute=False, bias=0,
@@ -266,7 +264,7 @@ class pudu:
 
         # Initial values
         sh = np.array(self.x).shape
[…]
