### 2023-09-15 #5645 @gkthiruvathukal
@enricgrau  Thank you for your submission to JOSS. We are currently facing a backlog of submissions in this track. Can you please follow up by looking at the [list of editors](https://joss.theoj.org/about#editorial_board)? When you reach this page, please scroll down to **Topic Editors**. Please note that I am not asking you to suggest reviewers at this stage. Please simply list the GitHub IDs of prospective topic editors **without** using the @.  Please reply to this thread by listing up to three GitHub IDs only (one per line).
