### 2023-09-24 #5645 @arfon
@aksholokhov @bbengfort @kmichael08 @hbaniecki  – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is `pudu`, A Python library for agnostic feature selection and explainability of Machine Learning classification and regression problems (https://github.com/pudu-py/pudu)

As authors of papers already published in JOSS on a similar topic, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out.

Many thanks
Arfon
