### 2023-10-13 pudu-py/pudu#1 @hbaniecki
[joss review] issues with examples
related to https://github.com/openjournals/joss-reviews/issues/5873
1. Examples in documentation should show outputs of the executed code, e.g. plots, could also use more description to guide users
2. Examples require data which is available in the repository (not clearly stated)
3. Example 1: Wrong code? No grouping in line 80
4. Example 3: wrong path `mnist_class.h5`
5. Example 4: the rainbow colour palette is not a good option for visualising feature importance
6. Example 5: error in imports `import perturbation as ptn`
7. Example 6: missing model `load_model('data/0_526_model_(80;79).h5')`
8. Example 7: wrong path `mnist_class.h5`
9. Running examples requires new package dependencies -- Installing them could be added as an option, e.g. with `pip install pudu[examples]`
