### 2023-10-13 pudu-py/pudu#2 @hbaniecki
[joss review] issues with documentation
related to https://github.com/openjournals/joss-reviews/issues/5873

1. Documentation https://pudu-py.github.io/pudu/definitions.html on Importance/Speed/Synergy seems sparse and has no references to relevant literature
2. Examples in documentation should show outputs of the executed code, e.g. plots, could also use more description to guide users
