### 2023-10-13 #5873 @hbaniecki
**Quality of writing**
- There is some vague language, e.g. "help make sense of machine learning results", "Easy plotting of the results"
- Acronyms "ML/XAI" are defined but later unused
- Typos in references: "Bhatt et al. (2020)](Belle & Papantonis, 2021)" "Bau2018"
