### 2023-10-14 pudu-py/pudu#3 @aksholokhov
[joss review] review from @aksholokhov
(related to [this](https://github.com/openjournals/joss-reviews/issues/5873) JOSS submission)

First, thank you for submitting your work to JOSS! I personally believe that pudu is a great package addressing a real need in its field and it may be eventually a JOSS paper. However, several issues need to be addressed for this to happen:

### Paper
**1. Relation to other works**
  The package is presented as an "understandable approach that can help the different stakeholders to better understand the AI algorithms" but it does not articulate what makes it more understandable than its competitors. Other libraries are mentioned (SHAP, LIME, GradCAM) as those that "can be difficult to implement, interpret, or understand for technical scientists
with little or no data science background." Is it a personal opinion of the authors or there is an indication of a consensus in the field (e.g. survey papers)? If former, then more justification is needed. For example, a table comparing all alternatives side by side with a reasonable quantification of effort to achieve what the authors need (e.g in lines of code, hours to implement, or just present/absent) would greatly improve the paper.

Also, the relationship between pudu and RELIEF remained unclear to me after reading the paper. Does pudu implement RELIEF and a set of wrappers around it that simplify the access to certain functionalities that are frequently needed in the fields of natural sciences? If so, then is it the only package that implements a sensitivity analysis with RELIEF? if yes, then mentioning it would greatly empower the statement of need. If no, then you need to summarize the differences between pudu and other packages that implement RELIEF.

**2. Evidence of academic interest**

I see that the package has over a thousand of downloads to date, which is great! Are there any works published that used pudu in their workflow? Mentioning them (or at least their total number to date) would be a great quantification of academic interest. Alternatively, mentioning example works that use the same analysis methodology (and thus could have benefited from pudu) would help.

### Code
Finally, the examples need a clean-up:

1. Some examples have import errors. E.g. `examples/example_pertubation.py` imports `perturbation as ptn` as a local import, whereas it is located in `pudu/perturbation` instead, which leads to import errors. It should be replaced with something like `from pudu import perturbation as ptn`.
2. Some examples import packages and modules that they don’t use. E.g. `examples/example_pertubations.py` imports `lime` which it does not use. Pudu also does not install it which leads to import check errors.
3. The examples should not assume their working directory, otherwise they fail when a project is opened at its root, for example when the whole project is opened in PyCharm. The examples fail because `data/` directory does not exist (it is in `examples/data`). The working directory should be inferred at run-time to ensure a smooth user experience.
4. Some examples try to read files that don’t exist. E.g. `examples_activation.py`  `tries to read data/0_526_model_(80;79).h5` which does not exist (at least not if the repository is cloned with git clone).
5. Finally, it would be great to have a paragraph describing (1) the context, (2) the data, (3) what the example aims to illustrate, and (4) what should I expect to see if this example runs correctly, in both the code file and in the documentation page.

Otherwise the package looks great!
