+++ examples/example_activations_1d.py
@@ -1,7 +1,6 @@
 import numpy as np
 from tensorflow import keras
 from keras.models import load_model
-import matplotlib.pyplot as plt
 import spectrapepper as spep
 
 from pudu import pudu, plots
@@ -18,7 +17,7 @@ y = [3, 3, 3] # these are all class 3
 y = keras.utils.to_categorical(y)
 
 # Load the model and test it
-model = load_model('data/0_526_model_(80;79).h5')
+model = load_model('data/1d_cnn.h5')
 score = model.evaluate(x, y, verbose=0)
 print("Test loss:", score[0], "| Test accuracy:", score[1])
 model.summary()
@@ -40,7 +39,7 @@ imp = pudu.pudu(x0, y0, cnn1d_prob, model)
 
 # First we check importance
 imp.importance(window=150, perturbation=ptn.Positive(delta=0.1))
-plots.plot(imp.x, imp.imp, axis=None, figsize=(10, 5), cmap='cool')
+plots.plot(imp.x, imp.imp, axis=None, figsize=(10, 4), cmap='cool')
 
 # Now we explore the unit activations observed in layer 4 (last conv. layer)
 # We only consider the top 0.5% of the values as activated with `p=0.005`.
@@ -55,17 +54,16 @@ plots.plot(imp.x, imp.lac, axis=None, figsize=(10, 4), cmap='cool',
 plots.feature_report(imp.fac, plot=True, print_report=False, show_top=10)
 plots.unit_report(imp.uac, plot=True, print_report=False, show_top=10)
 
-# To better visualize the coordinates, we can better check with 'preview'
-imp.preview(window=150, figsize=(10, 4), yticks=None, bold=True)
-
-# We can extract more infomration when using several data points and `relatable`. 
+# We can extract more information when using several data points and `relatable`. 
 # This will check, after applying `activations` to N images, what units activate
-# with what feature changes the most. Then, for N=3
-# Change the shape of the array
-x = x.reshape(3, 1, 14730, 1)
+# with what feature changes the most. This can be very computer-intensive. Here 
+# it only tests for 3 samples, but
+# you will need a larger N to get significant results. If you can use a GPU it is
+# highly advised to do so.
+x = x.reshape(3, 1, 14730, 1) # Change the shape of the array
 y = [np.argmax(i) for i in y]
 
 imp = pudu.pudu(x, y, cnn1d_prob, model) # we need to build another pudu
-imp.relatable(layer=2, slope=0, p=0.0025, window=150, perturbation=ptn.Positive(delta=0.1))
+imp.relatable(layer=2, slope=0, p=0.0025, window=200, perturbation=ptn.Positive(delta=0.1))
 idx, cnt, cor = imp.icc # this outputs the unit index, activation counts, and position of the feature
 plots.relate_report(idx, cnt, cor, plot=True, show_top=20, font_size=10, rot_ticks=45, sort=True)
