+++ docs/_sources/definitions.rst.txt
@@ -2,9 +2,24 @@
 Definitions
 ===========
 
+Spectroscopy studies the interaction between light and matter, and when analyzing this type of data, it is 
+all about change. Spectral data will show as an accumulation of peaks, and these will vary in shape, symetry, intensity, 
+position and convolutions. The most subtle changes can mean significant property differences in the 
+measured material, but also big changes could mean little (Vickerman & Gilmore, 2009). Thus, studying what these
+changes mean and how they occur is paramount. This is also true when using machine learning for spectroscopic data;
+change in spectral features remains the most important thing. Here we define 4 techniques included in **pudu** that
+evaluate ML models in terms of change, aiming to help scientists to further analyze their spectroscopic data beyond 
+the ML results.
+
+
 Importance
 ==========
-Let :math:`x \in X` be a 2-d array of dimensions :math:`h \times w`. Let :math:`P_M` be
+
+Importance quantifies the relevance of the features according to the changes in the prediction according to 
+defined sequential perturbations on the features. Thus, Importance is measured in probability or target value 
+difference for classification or regression problems, respectively.
+
+In a formal way, Let :math:`x \in X` be a 2-d array of dimensions :math:`h \times w`. Let :math:`P_M` be
 the probability function of the model :math:`M`. Then, :math:`P_M(x)` is the probability of :math:`x`
 to belong to a classification class according to the problem solved by :math:`M`. Considering 
 :math:`j \in J` the feature in position :math:`(h_j, w_j)` of :math:`x`, then the local importance 
@@ -24,13 +39,16 @@ in a model's classification probability according to change in the features.
 
 Speed
 =====
-Consider states of :math:`R` with different set parameters :math:`R_1, R_2,  ...`. As for
-importance for :math:`x`, the local importance of feature :math:`j` using the different 
-perturbaions would be :math:`LI_{j,1}, LI_{j,2}, ...`. Then, speed is the slope calculated
-according to the linear fit of the local importance points as
-:math:`\left( 1, LI_{j,1} \right), \left( 2, LI_{j,2} \right), ...`
+Speed quantifies how fast a prediction changes according to perturbations in the features. For this, 
+the Importance is calculated at different perturbation levels, and a line is fitted to the obtained 
+values and the slope is extracted as the Speed value.
 
-Then, the speed is how fast the imortance changes accoding to change in the feature, or
+This is better defined considering states of :math:`R` with different set parameters :math:`R_1, R_2,  ...`. 
+As for importance for :math:`x`, the local importance of feature :math:`j` using the different perturbaions 
+would be :math:`LI_{j,1}, LI_{j,2}, ...`. Then, speed is the slope calculated according to the linear fit of the 
+local importance points as :math:`\left( 1, LI_{j,1} \right), \left( 2, LI_{j,2} \right), ...`
+
+Then, the speed is how fast the Imortance changes accoding to change in the feature, or
 how sensitive it is. These can have positive or negative values, depending on the slope.
 A positive value means that a bigger change will produce a bigger change in the 
 prediciton. A negative value means that bigger changes produce smaller changes in the
@@ -39,18 +57,30 @@ prediction.
 
 Synergy
 =======
-Consider a feature :math:`j* \in J` and a distinct feature :math:`j \in J` from :math:`x_i`. Both are
+
+Peaks in spectral data can change at the same time as other peaks but their relationship can be difficult to 
+pinpoint and understand, especially in more complex mixtures and materials. Synergy helps to explore this 
[…]
