+++ docs/_sources/examples.rst.txt
@@ -1,57 +1,256 @@
 Examples
 --------
 
+These examples show common use cases and scenarios for **pudu**. Most examples use spectroscopic data, and some 
+of them complement with images to show 2-D capability. The later data is the mnist database and is downloadable 
+from the code itself. For the spectral data, parameters, and sample models please download the 
+`data <https://github.com/pudu-py/pudu/tree/main/examples/data>`_ folder included in the 
+`examples <https://github.com/pudu-py/pudu/tree/main/examples>`_ folder from the repository.
+
+
 Example 1: PC-LDA with Scikit-learn for classification of spectroscopic data
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+In this example, a PC-LDA cascaded algorithm built with Scikit-learn is evaluated using LIME and pudu. Both 
+methods show vastly different results. After setting up pudu allows to perform additional analysis and plotting 
+the results with single-line commands.
 
 .. literalinclude:: ../examples/example_pclda.py
     :linenos: 
 
 
+.. figure:: images/ex1_1.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex1_2.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex1_3.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex1_4.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex1_5.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+
+
 Example 2: Regression for predicition of open circuit voltage of solar cells with RBFN
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Regression can be an important method for spectrocopy. Here, a Radial Basis Funtion Network is trained with localreg
+to perform high-accuracy preduction of VOC with spectral areas calculated beforehand. pudu is used to analyze how 
+the regression behaves according to these areas.
+
 .. literalinclude:: ../examples/example_rbnf.py
     :language: python
+
+.. figure:: images/ex2_1.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex2_2.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex2_3.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex2_4.png
+   :scale: 50 %
+   :align: center
+   :alt: 
 
 
 Example 3: Handwritten numbers recognition using CNN with Keras and Mnist
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+This example showcases 2-D capabilities of pudu with a CNN for MNIST, alongise with LIME and GRAD-CAM visualizatinos.
+
 .. literalinclude:: ../examples/example_cnn_mnist.py
     :language: python
+
+.. figure:: images/ex3_1.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
+.. figure:: images/ex3_2.png
+   :scale: 50 %
+   :align: center
+   :alt: 
+
[…]
