+++ docs/_sources/introduction.rst.txt
@@ -29,18 +29,22 @@ Introduction to pudu
 * Conda-forge: https://anaconda.org/conda-forge/pudu
 * Free software: MIT license
 
-**pudu** is a Python package that helps intepret and explore the results of machinme learning algorythms. It does this by quantifying the change
-in probability according to the change in the features. This library works with any case that has a probability function, which is normally available in ``scikit-learn`` and ``keras`` methods, and works for both 1-d (vectrors) and 2-d problems (images). In order to see the exact procedure and format needed, please refer to the examples in the ``docs``.
+**pudu** is a Python package that helps interpret and explore the results of machine learning algorithms 
+with spectroscopic data. It does this by quantifying the change in the prediction according to the change 
+in the features. This library works with classification and regression problems with both 1-d and 2-d problems. 
 
 Features
 --------
 
 The following is a list of the main procedures that **pudu** package enables.
 
-- Importance: estimates the absolute or relative importance oif the features.
-- Speed: calculates how fast a preditci0on changes according to the changes in the features.
-- Synergy: tests teh synergy between features and the change in classification probability.
-- Easy plotting of the results from the above.
+The following is a list of the main features that **pudu** package enables.
+
+- Importance: measures the change in prediction according to perturbations in the features.
+- Speed: calculates how fast a prediction changes according to different perturbation levels.
+- Synergy: tests the synergy between features and the change in classification probability.
+- Re-activations: Evaluates how activations maps from CNNs change according to perturbations in the data.
+- Easy plotting with ample personalization options for all the cases above.
 
 Quickstart
 ----------
