+++ paper/OldVersions/paper.md
@@ -1,0 +1,77 @@
+---
+title: "pudu: A Python library for agnostic feature selection and explainability of Machine Learning classification and regression problems."
+tags:
+    - Python
+    - Machine Learning
+    - Explainability and intepretability
+    - Sensitivity Analysis
+    - Classification and regression
+authors:
+    - name: Enric Grau-Luque
+      orcid: 0000-0002-8357-5824
+      affiliation: "1"
+    - name: Ignacio Becerril-Romero
+      orcid: 0000-0002-7087-6097
+      affiliation: "1"
+    - name: Alejandro Perez-Rodriguez
+      orcid: 0000-0002-3634-1355
+      affiliation: "1, 2"
+    - name: Maxim Guc
+      orcid: 0000-0002-2072-9566
+      affiliation: "1"
+    - name: Victor Izquierdo-Roca
+      orcid: 0000-0002-5502-3133
+      affiliation: "1"
+affiliations:
+    - name: Catalonia Institute for Energy Research (IREC), Jardins de les Dones de Negre 1, 08930 Sant Adrià de Besòs, Spain
+      index: 1
+    - name: Departament d'Enginyeria Electrònica i Biomèdica, IN2UB, Universitat de Barcelona, C/ Martí i Franqués 1, 08028 Barcelona, Spain
+      index: 2
+date: 30 June 2023
+bibliography: paper.bib
+---
+
+
+# Statement of need
+
+Interpreting the results of machine learning (ML) algorithms has become an increasingly important topic in the field, as concerns about the lack of interpretability of these models have grown [@Burkart2021]. While many libraries and frameworks provide tools for making predictions and evaluating model performance, understanding how the predictions are being made and the factors that are driving them can be challenging. This lack of interpretability can strongly reduce the confidence in the results obtained with these models, particularly when they are being used to make important decisions.
+
+Natural sciences, in particular, have shown an increased interest in the use of machine learning algorithms due to their capability of analyzing large amounts of data in a fast and economic way, their generalized application, and easy access through several libraries and products. At the same time, explainable artificial intelligence (XAI) is being increasingly demanded in science since results obtained from AI analyses may lack scientific value if they cannot be properly interpreted, which can affect scientific consistency and diminish the significance of the results for different knowledge domains [@Roscher2020].
+
+One of the approaches to improve the interpretability of machine learning algorithms is through sensitivity analysis, which is based on systematically varying the input features and measuring the resulting change in the prediction. By comparing the predictions obtained with the original values of the features to those obtained with the modified values, it is possible to gain insight on the relations between the prediction and the different features. An example of the latter would be RELIEF [@Kira1992], a feature selection method that detects statistically significant features according to the changes that they produce in the target.
+
+Even though there are products and libraries available for explaining different types of algorithms such as SHAP [@Lundberg2017], LIME [@Ribeiro2016], or GradCAM [@Selvaraju2017], they can be difficult to implement, interpret, or understand for technical scientists with little or no data science background. As such, a tool with an easy and understandable approach that can help the different stakeholders to better understand the AI algorithms that they are employing in their data analyses can considerably increase the transparency, comprehensibility, and scientific impact of machine learning results in natural sciences and in other applications [@Bhatt2020][@Belle2021].
+
+# Overview
+
[…]
