+++ examples/example_masking.py
@@ -10,12 +10,12 @@ from pudu import masks as msk
 # This example shows the use of `mask` options available.
 
 # Load features (spectra) and targets (open circuit coltage, voc)
-features = spep.load('data/features.txt')
-targets = spep.load('data/targets.txt', transpose=True)[2]
+features = spep.load('examples/data/features.txt')
+targets = spep.load('examples/data/targets.txt', transpose=True)[2]
 
 # Load pre-trained LDA and PCA models
-lda = pickle.load(open('data/lda_model.sav', 'rb'))
-pca = pickle.load(open('data/pca_model.sav', 'rb'))
+lda = pickle.load(open('examples/data/lda_model.sav', 'rb'))
+pca = pickle.load(open('examples/data/pca_model.sav', 'rb'))
 
 ### PUDU ###
 # Select x (feature) and respective y (target)
@@ -37,5 +37,5 @@ imp = pudu.pudu(x, y, pf)
 masks = [msk.All(), msk.EveryOther(), msk.RandomMask()]
 m_names = ['All', 'EveryOther', 'RandomMask']
 for i,j in zip(masks, m_names):
-    imp.importance(window=100, absolute=False, mask=i)
+    imp.importance(window=25, absolute=False, mask=i)
     plots.plot(imp.x, imp.imp, title="Importance - "+j, yticks=[], font_size=15, cmap='winter')
