+++ paper/paper.bib
@@ -633,9 +633,8 @@ year = {2023}
 
 @article{Bhatt2020,
 abstract = {Explainable machine learning offers the potential to provide stake-holders with insights into model behavior by using various methods such as feature importance scores, counterfactual explanations, or influential training data. Yet there is little understanding of how organizations use these methods in practice. This study explores how organizations view and use explainability for stakeholder consumption. We find that, currently, the majority of deployments are not for end users affected by the model but rather for machine learning engineers, who use explainability to debug the model itself. There is thus a gap between explainability in practice and the goal of transparency, since explanations primarily serve internal stakeholders rather than external ones. Our study synthesizes the limitations of current explainability techniques that hamper their use for end users. To facilitate end user interaction, we develop a framework for establishing clear goals for explainability. We end by discussing concerns raised regarding explainability. CCS CONCEPTS • Human-centered computing; • Computing methodologies → Philosophical/theoretical foundations of artificial intelligence ; Machine learning; KEYWORDS machine learning, explainability, transparency, deployed systems, qualitative study ACM Reference Format:},
-author = {Bhatt, Umang and Xiang, Alice and Sharma, Shubham and Weller, Adrian and Taly, Ankur and Jia, Yunhan and Ghosh, Joydeep and Puri, Ruchir and Moura, Jos{\'{e}} M F and Eckersley, Peter},
+author = {Bhatt, Umang and Xiang, Alice and Sharma, Shubham and Weller, Adrian and Taly, Ankur and Jia, Yunhan and Ghosh, Joydeep and Puri, Ruchir and Moura, José M F and Eckersley, Peter},
 doi = {10.1145/3351095.3375624},
-file = {:C\:/Users/etgrau/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Bhatt et al. - 2020 - Explainable Machine Learning in Deployment.pdf:pdf},
 isbn = {9781450369367},
 keywords = {deployed systems,explainability,machine learning,qualitative study,transparency},
 mendeley-groups = {pudu},
