+++ paper/paper.md
@@ -34,9 +34,9 @@ bibliography: paper.bib
 
 # Statement of need
 
-Spectroscopic techniques (e.g. Raman, photoluminescence, reflectance, transmittance, X-ray fluorescence) are an important and widely used resource in different fields of science, such as photovoltaics [@Fonoll-Rubio2022][@Grau-Luque2021], cancer [@Bellisola2012], superconductors [@Fischer2007], polymers [@Easton2020], corrosion [@Haruna2023], forensics [@Bhatt2023], and environmental sciences [@Estefany2023], to name just a few. This is due to the versatile, non-destructive and fast acquisition nature that provides a wide range of material properties, such as composition, morphology, molecular structure, optical and electronic properties. As such, machine learning (ML) has been used to analyze spectral data for several years, elucidating their vast complexity, and uncovering further potential on the information contained within them [@Goodacre2003][@Luo2022]. Unfortunately, most of these ML analyses lack further interpretation of the derived results due to the complex nature of such algorithms. In this regard, interpreting the results of ML algorithms has become an increasingly important topic, as concerns about the lack of interpretability of these models have grown [@Burkart2021]. In natural sciences (like materials, physical, chemistry, etc.), as ML becomes more common, this concern has gained especial interest, since results obtained from ML analyses may lack scientific value if they cannot be properly interpreted, which can affect scientific consistency and strongly diminish the significance and confidence in the results, particularly when tackling scientific problems [@Roscher2020].
+Spectroscopic techniques (e.g. Raman, photoluminescence, reflectance, transmittance, X-ray fluorescence) are an important and widely used resource in different fields of science, such as photovoltaics [@Fonoll-Rubio2022] [@Grau-Luque2021], cancer [@Bellisola2012], superconductors [@Fischer2007], polymers [@Easton2020], corrosion [@Haruna2023], forensics [@Bhatt2023], and environmental sciences [@Estefany2023], to name just a few. This is due to the versatile, non-destructive and fast acquisition nature that provides a wide range of material properties, such as composition, morphology, molecular structure, optical and electronic properties. As such, machine learning (ML) has been used to analyze spectral data for several years, elucidating their vast complexity, and uncovering further potential on the information contained within them [@Goodacre2003] [@Luo2022]. Unfortunately, most of these ML analyses lack further interpretation of the derived results due to the complex nature of such algorithms. In this regard, interpreting the results of ML algorithms has become an increasingly important topic, as concerns about the lack of interpretability of these models have grown [@Burkart2021]. In natural sciences (like materials, physical, chemistry, etc.), as ML becomes more common, this concern has gained especial interest, since results obtained from ML analyses may lack scientific value if they cannot be properly interpreted, which can affect scientific consistency and strongly diminish the significance and confidence in the results, particularly when tackling scientific problems [@Roscher2020].
 
[…]
