### 2023-10-31 pudu-py/pudu#2 @enricgrau
@hbaniecki Thank you for opening this issue. Regarding your observations:

1. We have improved the `definitions` section. However, as these are methodologies that we are introducing, we have struggled to add additional references to them. We hope that considering the new focus of the manuscript and library regarding spectroscopic data those definitions make more sense. Despite this, we are happy to look deeper into the matter if you believe this is paramount for the publication.
2. We now show an improved `examples` section adding the output images and more context to each example.
