### 2023-10-31 pudu-py/pudu#1 @enricgrau
@hbaniecki Thank you for pointing out all these issues.
1. We have improved this section by adding the output images and more context to each example.
2. We now explicitly state at the beginning of the section that data has to be downloaded.
3. The example was fixed.
4. The path is fixed.
5. We have changed the `jet` palette.
6. Import fixed.
7. The model is now included in the repository.
8. Path fixed.
9. We now include an `examples_requierments.txt` in the `examples` folder, indicating that these additional libraries can be installed from there. We can change this approach if you suggest otherwise.
