### 2023-10-31 pudu-py/pudu#2 @hbaniecki
Thanks for clarifying that this methodology is introduced here
