### 2023-11-03 pudu-py/pudu#4 @enricgrau
Response to comments from @hbaniecki
> Apart from minor issues with [documentation](https://github.com/pudu-py/pudu/issues/2) and [examples](https://github.com/pudu-py/pudu/issues/1), I have the following major concerns about this contribution:
>
> 1. **Substantial scholarly effort is (currently) not evident.** It is a relatively young software (4k downloads), there is no user base (no issues, no GitHub stars, no blogs or promotion of the software?), and no clear applications (e.g. no citations). With similar software existing already, {pudu} is not likely to be cited unless the software/paper clearly highlights its original purpose, e.g. an original subset of methods implemented, accessible visualisation, better API, unique audience.
> 2. **Statement of need is unclear.** I have a hard time understanding the motivation for developing this software. In its current state, the paper's title and description seem too generic. Should "sensitivity analysis" be highlighted in the title?
>
>    * A clear use-case for this software would help, e.g. the paper mentions the RELIEF method but does not elaborate on applications/usefulness of the algorithm to show that the {pudu} package is indeed needed in situation X or for user Y.
>    * The terminology of "Importance/speed/synergy/reactivation" is unclear and could be better described in the paper.
>    * Adding a figure to the paper could help illustrate the statement of need.
> 3. **State of the field is missing.** There is no description of how this software compares to other similar packages. Consider the attached non-exhaustive list of software for explainability and feature selection. Relating {pudu} to other software could partially alleviate concerns mentioned in 1. & 2., i.e. show the added value of {pudu} over the already available solutions, motivate the need for such new package, highlight the target audience etc.
>
> I am open to discussion and hope the software paper can be improved to clearly state the motivation and effort.
>
> **References (non-exhaustive list)**
>
> * Molnar et al. iml: An R package for Interpretable Machine Learning. JOSS 2018
> * Alber et al. iNNvestigate Neural Networks! JMLR 2019
> * Kokhlikyan et al. Captum: A unified and generic model interpretability library for PyTorch. arXiv:2009.07896 2020
> * Arya et al. AI Explainability 360: An Extensible Toolkit for Understanding Data and Machine Learning Models. JMLR 2020
> * (ours) dalex: Responsible Machine Learning with Interactive Explainability and Fairness in Python. JMLR 2021
> * Klaise et al. Alibi Explain: Algorithms for Explaining Machine Learning Models. JMLR 2021
> * Li et al. InterpretDL: Explaining Deep Models in PaddlePaddle. JMLR 2022
> * Zhu et al. abess: A Fast Best-Subset Selection Library in Python and R. JMLR 2022

> **Quality of writing**
>
> * There is some vague language, e.g. "help make sense of machine learning results", "Easy plotting of the results"
> * Acronyms "ML/XAI" are defined but later unused
> * Typos in references: "Bhatt et al. (2020)](Belle & Papantonis, 2021)" "Bau2018"
