### 2023-11-03 pudu-py/pudu#4 @enricgrau
Regarding the review of @hbaniecki in https://github.com/openjournals/joss-reviews/issues/5873#issuecomment-1761368406

Thank you for your comprehensive and detailed comments. We appreciate the time you took to review our work and constructively provide your concerns. We have thoroughly revised the manuscript following your suggestions and concerns, and we believe that now shows to be much improved. The article does appear to be quite different from the original one. We intentionally removed and changed sections and several phrases and paragraphs to regain focus on the original inspiration for the library. We hope that now the article shows the motivation, opportunity, and application clearly.

Here we respond point by point:

1. We acknowledge that `pudu` is a relatively young software with limited visibility in its current state. We hope that the changes that we have made in addition to a potential publication, can help the library to gain traction with the target audience. The library comes from a very specific concern regarding research with spectroscopic techniques, but we tried to present it more generically and broadly to expand the target user. However, it is clear now to us, due to your observations, that we didn’t make the mark on building a solid case for such a broad scope. With this in mind, we have decided to narrow the reach of the library to its original motivation regarding spectroscopic data. This is now reflected in the manuscript, and we hope to be clearer on why we believe that pudu can contribute to the field.
2. We recognize that the motivation behind the development of pudu needs to be more explicitly communicated. As mentioned, the initial motivation is very specific, and we tried to make its motivation more generalized to hopefully capture more interest. To this end, we restructured this section to highlight the specific gaps and challenges with existing tools. We include relevant survey articles that observe problems with explainability methods and libraries and also better explain why `pudu` fits the gap between spectroscopy-related research and explainable methods.

Regarding the relationship with RELIEF: Our intention with pudu was to be a wrapper for the basic principles of RELIEF and to offer extended functionality that addresses the challenges in our field of research. After much deliberation, we decided to remove the reference to RELIEF since it causes more confusion than clarity, and the reference offers no value regardless.

We agree that the terminology of Importance/speed/synergy/reactivation are vague and unclear in the paper. The original idea was for the reader to complement the paper with the documentation where these are better explained. We realize now that we failed to achieve this. Thus, we have improved these explanations in the paper to better define these for the reader in a more clear and accessible way.

Following your advice, to further aid understanding we added a figure to visually represent the primary motivations and benefits of using pudu, giving readers a clear snapshot of its utility.

3. Considering that we have restructured the manuscript to focus on spectroscopy, we believe that now it better shows the context and need on why pudu is beneficial for spectroscopy, using relevant references to support this. We hope this will offer readers a clear understanding of pudu's added value.

Quality of writing:
- We worked on identifying this and other vague statements and transformed them into concrete ones. We are happy to keep improving such vague phrases if needed.
- We have improved the use of acronyms and only making definitions if needed.
- Typos with references were fixed. The problem was spacing between brackets.
