### 2023-11-03 pudu-py/pudu#3 @enricgrau
Thank you @aksholokhov for your thoughtful review. We truly appreciate your comprehensive feedback and insight. After serious consideration, we believe that addressing the concerns you raised has strengthened our manuscript. We have thoroughly reviewed the paper in its entirety with your suggestions and concerns in mind, and we believe we have improved it considerably. The article does appear to be quite different from the original one. We have intentionally made several changes in the sections and deleted many other phrases and paragraphs so we can focus on the main and original goal of the library, which is to aid in the analysis of spectroscopy using ML. We hope that now the article shows the motivation, opportunity, and application clearly and matches your expectations for a JOSS publication.

Here are our responses to each of the points:

**Paper**
1. Indeed, we have ample anecdotal evidence to say that such methodologies are difficult to use. Additionally, we are aware of survey studies tackling this and other issues. As such, we rephrase that paragraph in the revised version referencing relevant literature about these problems. Additionally, we recognize the importance of distinguishing `pudu` from other libraries and methods, thus we offer better context on why `pudu` can find to be useful for researchers using spectrocopy and ML.

As for the relationship between `pudu` and the **RELIEF** method, after extended discussions, we have decided to exclude this reference. In retrospect, we now realize that it only introduces confusion to the goal of the library and offers little value to the overall work. Instead, we focus on the specific method used by `pudu` and its relevance to the field.

2. As `pudu` is relatively new and only has been in a real functional and usable form after version 0.3.0, we do not expect for peers to use it in current publications. However, we believe that with the changes and insights provided in this submission review and a potential publication, the library will get increased visibility and interest within the target audience.

**Code**
1. We have fixed these import errors in all examples.
2. We have revised these imports on the examples and now only include the ones in use. Additionally, we include a `examples_requierments.txt` file with the additional libraries needed to run all the examples.
3. We have changed the directory of all data imports to `examples/data`.
4. We now include all the files referenced in the examples.
5. We now include more context for each of the examples following your points, also with the output images.
