### 2023-11-03 pudu-py/pudu#4 @enricgrau
@hbaniecki Please find the new manuscript in the library's repository [here](https://github.com/pudu-py/pudu/blob/main/paper/paper.pdf) in the meantime. We are not able to re-render with the `editorialbot` from JOSS for some reason, but all the changes that will shown once this is solved are reflected in this preview generated from the new `paper.md`.
