### 2023-11-03 #5873 @enricgrau
@arfon We have responded to @aksholokhov and @hbaniecki in issues https://github.com/pudu-py/pudu/issues/4 and https://github.com/pudu-py/pudu/issues/3. At the moment, we ask to check the preview from the new and revised `paper.md` found in the repository [here](https://github.com/pudu-py/pudu/blob/main/paper/paper.pdf). The new `pdf` should be generated from this same file so we hope this causes no problems in the review process. Thank you all again for your valuable time and have a great weekend! 😄
