### 2023-11-06 24ba369
delete paper.md
