+++ paper/OldVersions/paper.bib
@@ -1,726 +1,0 @@
-@article{Goodacre2003,
-abstract = {Whole organism or tissue profiling by vibrational spectroscopy produces vast amounts of seemingly unintelligible data. However, the characterisation of the biological system under scrutiny is generally possible only in combination with modern supervised machine learning techniques, such as artificial neural networks (ANNs). Nevertheless, the interpretation of the calibration models from ANNs is often very difficult, and the information in terms of which vibrational modes in the infrared or Raman spectra are important is not readily available. ANNs are often perceived as 'black box' approaches to modelling spectra, and to allow the deconvolution of complex hyperspectral data it is necessary to develop a system that itself produces 'rules' that are readily comprehensible. Evolutionary computation, and in particular genetic programming (GP), is an ideal method to achieve this. An example of how GP can be used for Fourier transform infrared (FT-IR) image analysis is presented, and is compared with images produced by principal components analysis (PCA), discriminant function analysis (DFA) and partial least squares (PLS) regression. {\textcopyright} 2003 Elsevier B.V. All rights reserved.},
-author = {Goodacre, Royston},
-doi = {10.1016/S0924-2031(03)00045-6},
-issn = {0924-2031},
-journal = {Vibrational Spectroscopy},
-keywords = {Artificial neural networks,FT-IR,Genetic programming},
-month = {aug},
-number = {1},
-pages = {33--45},
-publisher = {Elsevier},
-title = {{Explanatory analysis of spectroscopic data using machine learning of simple, interpretable rules}},
-volume = {32},
-year = {2003}
-}
-
-
-@article{Luo2022,
-abstract = {Raman spectroscopy (RS) is a spectroscopic method which indirectly measures the vibrational states within samples. This information on vibrational states can be utilized as spectroscopic fingerprints of the sample, which, subsequently, can be used in a wide range of application scenarios to determine the chemical composition of the sample without altering it, or to predict a sample property, such as the disease state of patients. These two examples are only a small portion of the application scenarios, which range from biomedical diagnostics to material science questions. However, the Raman signal is weak and due to the label-free character of RS, the Raman data is untargeted. Therefore, the analysis of Raman spectra is challenging and machine learning based chemometric models are needed. As a subset of representation learning algorithms, deep learning (DL) has had great success in data science for the analysis of Raman spectra and photonic data in general. In this review, recent developments of DL algorithms for Raman spectroscopy and the current challenges in the application of these algorithms will be discussed.},
-author = {Luo, Ruihao and Popp, Juergen and Bocklitz, Thomas},
-doi = {10.3390/analytica3030020},
-file = {:C\:/Users/Enric/Downloads/analytica-03-00020.pdf:pdf},
-issn = {26734532},
-journal = {Analytica},
-keywords = {Raman spectroscopy,chemometrics,deep learning,machine learning},
-number = {3},
-pages = {287--301},
-title = {{Deep Learning for Raman Spectroscopy: A Review}},
-volume = {3},
-year = {2022}
-}
-
-
-@article{Easton2020,
[…]
