### 2023-11-06 #5873 @enricgrau
> > @arfon We have modified the paper.md and asked the editorialbot to re-generate the pdf but it is rendering the same old version. Can you help us with this? Shall we wait more time to re-generate? Thank you!
>
> I think this might be happening as you now have two `paper.md` files in the repository. @editorialbot will simply compile the first one it finds. Could you delete the one you do not want to be compiled?

That did the trick. Thank you!
