### 2023-11-06 pudu-py/pudu#4 @hbaniecki
Hi @enricgrau, thank you for the exhaustive answer. I re-read the revised version of the manuscript, which is much more precise now. I better understand the motivation behind this software and recognize the effort to introduce explainability into spectroscopy-related research.

In my opinion, it would still be useful to reference any other related explainability toolboxes (shap/lime/gradcam could be considered only an implementation of a particular method) so as to let the reader know about the benefits of using **pudu** in spectroscopic problems.

Figure 1 is mentioned only briefly in the text. Adding more context regarding the predictive task/data is needed for it to be understandable by a broader audience.

other minor:
- Should the {spectrapepper} software (JOSS 2021) be discussed as related work?
- L70: name **pudu** can be bolded to be consistent
- L71: there is a typo 2x comma
- title: probably shouldn't end with a dot as it is later wrongly propagated to BibTeX, GScholar etc.

I believe the above remaining comments are minor/subjective enough to recommend the paper's acceptance.
