### 2023-11-06 #5873 @hbaniecki
Hi, I believe authors did a good job at improving the software/paper. My remaining comments are minor (see https://github.com/pudu-py/pudu/issues/4).

I can recommend acceptence of the {pudu} software paper to JOSS.
