### 2023-11-19 #5873 @aksholokhov
@arfon The authors addressed my feedback in full and I can recommend the acceptance of the {pudu} paper to JOSS as well.
