### 2023-11-20 66a5534
last version of paper with minor changes and badges
