### 2023-12-09 #5873 @arfon
@enricgrau – could you check your references in your BibTeX file please? It looks like there are duplicate entries for `Bhatt2020` and `Belle2021`
