### 2023-12-10 b05ec61
fix bibtext entries
