+++ paper/paper.bib
@@ -120,7 +120,7 @@ abstract = {Tunneling spectroscopy has played a central role in the experimental
 archivePrefix = {arXiv},
 arxivId = {cond-mat/0610672},
 author = {Fischer, {\O}ystein and Kugler, Martin and Maggio-Aprile, Ivan and Berthod, Christophe and Renner, Christoph},
-doi = {10.1103/REVMODPHYS.79.353/FIGURES/62/MEDIUM},
+doi = {10.1103/REVMODPHYS.79.353},
 eprint = {0610672},
 issn = {00346861},
 journal = {Reviews of Modern Physics},
@@ -344,52 +344,9 @@ volume = {8},
 year = {2020}
 }
 
-@article{Bhatt2020,
-abstract = {Explainable machine learning offers the potential to provide stake-holders with
-insights into model behavior by using various methods such as feature importance scores,
-counterfactual explanations, or influential training data. Yet there is little understanding of
-how organizations use these methods in practice. This study explores how organizations view and
-use explainability for stakeholder consumption. We find that, currently, the majority of
-deployments are not for end users affected by the model but rather for machine learning
-engineers, who use explainability to debug the model itself. There is thus a gap between
-explainability in practice and the goal of transparency, since explanations primarily serve
-internal stakeholders rather than external ones. Our study synthesizes the limitations of
-current explainability techniques that hamper their use for end users. To facilitate end user
-interaction, we develop a framework for establishing clear goals for explainability. We end by
-discussing concerns raised regarding explainability.},
-author = {Bhatt, Umang and Xiang, Alice and Sharma, Shubham and Weller, Adrian and Taly, Ankur and Jia, Yunhan and Ghosh, Joydeep and Puri, Ruchir and Moura, Jos{\'{e}} M F and Eckersley, Peter},
-doi = {10.1145/3351095.3375624},
-isbn = {9781450369367},
-keywords = {deployed systems,explainability,machine learning,qualitative study,transparency},
-mendeley-groups = {pudu},
-title = {{Explainable Machine Learning in Deployment}},
-url = {https://doi.org/10.1145/3351095.3375624},
-year = {2020}
-}
 
 @article{Belle2021,
-abstract = {Artificial intelligence (AI) provides many opportunities to improve private and
-public life. Discovering patterns and structures in large troves of data in an automated manner
-is a core component of data science, and currently drives applications in diverse areas such as
-computational biology, law and finance. However, such a highly positive impact is coupled with
-a significant challenge: how do we understand the decisions suggested by these systems in order
-that we can trust them? In this report, we focus specifically on data-driven methods—machine
-learning (ML) and pattern recognition models in particular—so as to survey and distill the
-results and observations from the literature. The purpose of this report can be especially
-appreciated by noting that ML models are increasingly deployed in a wide range of businesses.
-However, with the increasing prevalence and complexity of methods, business stakeholders in the
-very least have a growing number of concerns about the drawbacks of models, data-specific
-biases, and so on. Analogously, data science practitioners are often not aware about approaches
-emerging from the academic literature or may struggle to appreciate the differences between
-different methods, so end up using industry standards such as SHAP. Here, we have undertaken a
-survey to help industry practitioners (but also data scientists more broadly) understand the
-field of explainable machine learning better and apply the right tools. Our latter sections
-build a narrative around a putative data scientist, and discuss how she might go about
-explaining her models by asking the right questions. From an organization viewpoint, after
-motivating the area broadly, we discuss the main developments, including the principles that
-allow us to study transparent models vs. opaque models, as well as model-specific or
-model-agnostic post-hoc explainability approaches. We also briefly reflect on deep learning
-models, and conclude with a discussion about future research directions.},
[…]
