### 2023-12-10 9c9ed28
fix bibtext entries, final
