+++ paper/paper.bib
@@ -34,14 +34,14 @@ year = {2022}
 @article{Easton2020,
 abstract = {XPS is widely used to identify and quantify the elements present at the surface of polymeric materials. The energy distribution of photoelectrons emitted from these elements contains information about their chemical state, potentially allowing the analyst to identify and quantify specific functional groups. These functional groups may originate from the synthesis and processing of the polymers, from postsynthetic modifications such as surface grafting, or indeed may be unrelated to the polymer (additives and contaminants). Extracting reliable and meaningful information from XPS data is not trivial and relies on careful and appropriate experimentation, including experimental design, sample preparation, data collection, data processing, and data interpretation. Here, the authors outline some of these challenges when performing XPS analysis of polymers and provide practical examples to follow. This guide will cover all relevant aspects over the course of a typical experiment, including tips and considerations when designing the experiment, sample preparation, charge neutralization, x-ray induced sample damage, depth profiling, data analysis and interpretation, and, finally, reporting of results. Many of these topics are more widely applicable to insulating organic materials, and the recommendations of this guide will help to ensure that data is collected and interpreted using current best practices.},
 author = {Easton, Christopher D. and Kinnear, Calum and McArthur, Sally L. and Gengenbach, Thomas R.},
-doi = {10.1116/1.5140587/247679},
+doi = {10.1116/1.5140587},
 issn = {0734-2101},
 journal = {Journal of Vacuum Science & Technology A: Vacuum, Surfaces, and Films},
 month = {mar},
 number = {2},
 publisher = {American Vacuum Society},
 title = {{Practical guides for x-ray photoelectron spectroscopy: Analysis of polymers}},
-url = {/avs/jva/article/38/2/023207/247679/Practical-guides-for-x-ray-photoelectron},
+url = {https://pubs.aip.org/avs/jva/article-abstract/38/2/023207/247679/Practical-guides-for-x-ray-photoelectron?redirectedFrom=fulltext},
 volume = {38},
 year = {2020}
 }
