### 2023-12-10 #5873 @enricgrau
@arfon No problem :) I fixed the doi's and also deleted de duplicate entries. Thank you!
