+++ README.md
@@ -11,7 +11,7 @@
 [![image](https://img.shields.io/badge/stackoverflow-Ask%20a%20question-brown?logo=stackoverflow&logoWidth=18&logoColor=white)](https://stackoverflow.com/questions/tagged/pudu)
 [![status](https://joss.theoj.org/papers/cacb5b6520209b0c940bf46638df251d/status.svg)](https://joss.theoj.org/papers/cacb5b6520209b0c940bf46638df251d)
 
-**A Python library for explainability of machine learinng algorithms in an agnostic, deterministic, and simple way.**
+**A Python library for explainability of machine learning algorithms for spectroscopic data in an agnostic, deterministic, and simple way.**
 
 * GitHub repo: https://github.com/pudu-py/pudu
 * Documentation: https://pudu-py.github.io/pudu
